LayoutTests/imported/w3c/web-platform-tests/digital-credentials/get.tentative.https-expected.txt



PASS Calling navigator.credentials.get() without an digital member same origin.
PASS navigator.credentials.get() API rejects if there are no credential request.
PASS navigator.credentials.get() API rejects if there are no credential request for same-origin iframe.
PASS navigator.credentials.get() API rejects if there are no credential request in cross-origin iframe.
PASS navigator.credentials.get() promise is rejected if called with an aborted controller.
PASS navigator.credentials.get() promise is rejected if called with an aborted controller in same-origin iframe.
PASS navigator.credentials.get() promise is rejected if called with an aborted signal in cross-origin iframe.
PASS navigator.credentials.get() promise is rejected if abort controller is aborted after call to get().
FAIL navigator.credentials.get() promise is rejected if abort controller is aborted after call to get() in cross-origin iframe. assert_equals: expected "AbortError" but got "NotAllowedError"
PASS Mediation is required to get a DigitalCredential.
